Invert tagging invocation logic (#50)
* Fix SSH timeout when starting AWS instances

Since stopped instance do not have a public DNS name, the SSH
connections to them could not be established. The problem is solved by
acquiring the remote facts after the instances are started.

* Add osmetrics to cassandra cluster

Each cassandra node is provisioned with osmetrics. Updated the osmetrics
role to support both deb and rpm distributions.