AWS volumes setup (#59)
* Add volumes to monitoring machine

The services could not function due to the low disk space on the default
instance. It is resolved by adding a 30GB volume to be used by the
system, and additional 100GB volume for storing InfluxDB and ES data.

* Downgrade diagnostics due to metrics module issues

* Enable the filesystem task

Pre-task was not started at all due to the improper use of "when"
keyword. Besides fixing the startup play, all grafana JSON templates
have been fixed as well.

* Change Influx access to 'proxy'